requirements.txt
paddlepaddle
numpy
opencv-python
Pillow